Refactor pass #1 (unfinished).
This refactor introduces a number of important C++11 changes, it also
ensures consistency in the API with methods defined in the large
majority of classes in the library.

The refactor also works hard to be const correct, to the extreme I
prefer, so that method arguments are always const even for simple times
because of the number of times I've let functions grow big because I
never had time to refactor and break them up a bit.

Also made sure that every header had the same license blurb at the top
of the file, updated for 2020.